+++ lib/src/migrators/module.dart
@@ -252,7 +252,15 @@ class _ModuleMigrationVisitor extends MigrationVisitor {
           _shouldForward(declaration.name, forImportOnly: true) && !private;
       if (!visibleAtEntrypoint && !shouldBeVisible) continue;
 
-      var url = visibleAtEntrypoint ? entrypoint : declaration.sourceUrl;
+      Uri url;
+      if (visibleAtEntrypoint) {
+        url = entrypoint;
+      } else if (declaration.isImportOnly) {
+        url = declaration.importOnlyUrl;
+      } else {
+        url = declaration.sourceUrl;
+      }
+
       var prefix = '';
       if (renamedMembers.containsKey(declaration) ||
           (visibleAtEntrypoint && _startsWithPrefix(declaration.name))) {