+++ lib/src/migrators/module/member_declaration.dart
@@ -44,9 +44,16 @@ class MemberDeclaration<T extends SassNode> {
   /// This is `null` when [forward] is.
   final Uri forwardedUrl;
 
+  /// The canonical URL of the import-only stylesheet in which this member was
+  /// forwarded.
+  ///
+  /// This is `null` when this member wasn't loaded through an import-only
+  /// stylesheet.
+  final Uri importOnlyUrl;
+
   /// True if this declaration is the result of a `@forward` rule within an
   /// import-only stylesheet.
-  final bool isImportOnly;
+  bool get isImportOnly => importOnlyUrl != null;
 
   /// Constructs a MemberDefinition for [member], which must be a
   /// [VariableDeclaration], [Argument], [MixinRule], or [FunctionRule].
@@ -63,7 +70,7 @@ class MemberDeclaration<T extends SassNode> {
         sourceUrl = member.span.sourceUrl,
         forward = null,
         forwardedUrl = null,
-        isImportOnly = false;
+        importOnlyUrl = null;
 
   /// Constructs a forwarded MemberDefinition of [forwarding] based on
   /// [forward].
@@ -71,10 +78,12 @@ class MemberDeclaration<T extends SassNode> {
       MemberDeclaration forwarding, this.forward, this.forwardedUrl)
       : member = forwarding.member,
         name = '${forward.prefix ?? ""}${forwarding.name}',
-        isImportOnly = isImportOnlyFile(forward.span.sourceUrl),
         sourceUrl = isImportOnlyFile(forward.span.sourceUrl)
             ? forwarding.sourceUrl
-            : forward.span.sourceUrl;
+            : forward.span.sourceUrl,
+        importOnlyUrl = isImportOnlyFile(forward.span.sourceUrl)
+            ? forward.span.sourceUrl
+            : null;
 
   operator ==(other) =>
       other is MemberDeclaration &&